Adding new module for metadata calculation, first draft of code, cleaning up annotation src
