+++ src/matUtils/uncertainty.hpp
@@ -1,0 +1,8 @@
+#include "common.hpp"
+
+std::vector<MAT::Node*> get_common_nodes (std::vector<std::vector<MAT::Node*>> nodepaths);
+std::vector<float> get_all_distances(MAT::Node* target, std::vector<std::vector<MAT::Node*>> paths);
+size_t get_neighborhood_size(std::vector<MAT::Node*> nodes, MAT::Tree* T);
+void findEPPs (MAT::Tree Tobj, std::string sample_file, std::string fepps, std::string fneigh);
+po::variables_map parse_uncertainty_command(po::parsed_options parsed);
+void uncertainty_main(po::parsed_options parsed);
